% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{Backtest a trading strategy}
\usage{
backtest(data, strat, ...)
}
\arguments{
\item{data}{Data to be passed on to the Python backtest function}

\item{strat}{Character string indicating the strategy to backtest}

\item{...}{Paramters to be passed on to the Python backtest function}
}
\description{
Backtest a trading strategy
}
