% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-crypto-data.R
\name{get_crypto_data}
\alias{get_crypto_data}
\title{Returns price data for cypto symbol through Binance API}
\usage{
get_crypto_data(symbol, start_date, end_date)
}
\arguments{
\item{symbol}{A string indicating the ticker symbol of the cryptocurrency in
the Binance exchange}

\item{start_date}{A string indicating a date in the YYYY-mm-dd format, serves
as the start date of the period to get crypto data}

\item{end_date}{A string indicating a date in the YYYY-mm-dd format, serves
as the end date of the period to get crypto data}
}
\value{
A tibble
}
\description{
Returns price data for cypto symbol through Binance API
}
