% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pse.R
\name{get_pse_data}
\alias{get_pse_data}
\title{Returns pricing data for a specified stock}
\usage{
get_pse_data(sym, s_date, e_date)
}
\arguments{
\item{sym}{A string indicating the symbol of the stock in the PSE. For more
details, you can refer to this \href{https://www.pesobility.com/stock}{link}.}

\item{s_date}{A string indicating a date in the YYYY-mm-dd format, serves
as the start date of the period to get stock data}

\item{e_date}{A string indicating a date in the YYYY-mm-dd format, serves
as the end date of the period to get stock data}
}
\value{
A tibble, with the following columns:
\itemize{
\item symbol: The ticker symbol of the stock
\item dt: The date for the closing price of the stock
\item name: The name of the company represented by the stock ticker
\item currency: The currency of the closing price of the stock
\item close: The closing price of the stock at the given date, dt
\item percent_change: The percentage day change of the stock
\item volume: The total value of shares traded of the stock at dt
}
}
\description{
Returns pricing data for a specified stock
}
