% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastquant-install.R
\name{install_fastquant_source}
\alias{install_fastquant_source}
\title{Install fastquant from directory}
\usage{
install_fastquant_source()
}
\description{
Install fastquant from directory
}
\keyword{Internal}
