% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastquant-use.R
\name{use_fastquant_py}
\alias{use_fastquant_py}
\title{Import fastquant python module}
\usage{
use_fastquant_py()
}
\value{
A Python module object
}
\description{
Access backtesting functionality in R using the Python package. R does not
have native support for backtesting according to the backtrader framework.
}
