% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-use.R
\name{use_python}
\alias{use_python}
\title{Declare custom python version}
\usage{
use_python(python, required = FALSE)
}
\arguments{
\item{python}{Path to Python binary}

\item{required}{Is this version of Python required? If TRUE then an error
occurs if it's not located. Otherwise, the version is taken as a hint only
and scanning for other versions will still proceed.}
}
\description{
Declare custom python version
}
